+++ test/test-cabal.sh
@@ -1,0 +1,9 @@
+#!/usr/bin/env bash
+
+set -o xtrace
+set -o errexit
+
+cabal new-update
+
+cd /mnt
+cabal new-build example